requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
PyJWT==2.10.1
Werkzeug==3.1.3
pymongo[srv]
Flask-PyMongo
python-dotenv
pymongo
flask_jwt_extended
